change size_t to u64 where its serialized to disk
On 64bit windows and linux, size_t is equal to an u64.
However, Apple toolchains seem to treat it as a distinct type,
and fail to find a specialization for FileStream::read()/::write().

By setting these types explicitly to u64,
- we do not break existing savegames on 64bit Windows/Linux
- also make savegames from 32bit builds compatible (but break
  existing ones, but I doubt that these exist).

+++ TheForceEngine/TFE_Game/igame.cpp
@@ -17,7 +17,7 @@ MemoryRegion* s_levelRegion = nullptr;
 void displayMemoryUsage(const ConsoleArgList& args)
 {
 	char res[256];
-	size_t blockCount, blockSize;
+	u64 blockCount, blockSize;
 	region_getBlockInfo(s_gameRegion, &blockCount, &blockSize);
 	TFE_Console::addToHistory("-------------------------------------------------------------------");
 	TFE_Console::addToHistory("Region   | Memory Used | Current Capacity | Block Count | BlockSize");
